requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
Babel==2.8.0
black==19.10b0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
django-cors-headers==3.2.1
django-phonenumber-field==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.8
oauth2client==4.1.3
oauthlib==3.1.0
pathspec==0.7.0
phonenumbers==8.11.4
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
